move it to utils
